refactor: move all files into packages/*
