--- example/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "react-component-catalog-example",
-  "version": "0.0.1",
-  "description": "A working example for react-component-catalog",
-  "main": "server/index.js",
-  "private": true,
-  "scripts": {
-    "build": "NODE_ENV=production BABEL_ENV=production webpack -p",
-    "prebuild": "rimraf dist",
-    "start": "node .",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "watch": "npm run watch-client | npm run watch-server",
-    "watch-client": "webpack --watch",
-    "watch-server": "nodemon ."
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/natterstefan/react-component-catalog.git"
-  },
-  "keywords": [
-    "react",
-    "components",
-    "registry"
-  ],
-  "author": "Stefan Natter <stefan@natter.at>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/natterstefan/react-component-catalog/issues"
-  },
-  "homepage": "https://github.com/natterstefan/react-component-catalog#readme",
-  "peerDependencies": {
-    "babel-core": "6.x || ^7.0.0-bridge.0"
-  },
-  "dependencies": {
-    "@babel/register": "7.8.3",
-    "@hapi/hapi": "^19.1.1",
-    "@hapi/inert": "^6.0.1",
-    "dotenv": "8.2.0",
-    "react": "16.8.6",
-    "react-component-catalog": "file:..",
-    "react-dom": "16.8.6"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.8.4",
-    "@babel/core": "7.8.4",
-    "@babel/plugin-proposal-class-properties": "7.8.3",
-    "@babel/plugin-proposal-object-rest-spread": "7.8.3",
-    "@babel/polyfill": "7.8.3",
-    "@babel/preset-env": "7.8.4",
-    "@babel/preset-react": "7.8.3",
-    "awesome-typescript-loader": "5.2.1",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-plugin-dev-expression": "0.2.2",
-    "nodemon": "2.0.2",
-    "rimraf": "3.0.2",
-    "source-map-loader": "0.2.4",
-    "typescript": "3.7.5",
-    "webpack": "4.41.6",
-    "webpack-cli": "3.3.11"
-  }
-}
