+++ packages/example/babel.config.js
@@ -1,13 +1,6 @@
 module.exports = {
   presets: [
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
-      },
-    ],
+    ['@babel/preset-env', { useBuiltIns: 'entry', corejs: '3.6' }],
     '@babel/preset-react',
   ],
   plugins: [
