+++ packages/example/src/server/index.js
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
+/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable no-console */
+require('@babel/register')({
+  babelrc: true,
+})
+
+require('./server')
