+++ packages/example/src/server/server.js
@@ -1,9 +1,12 @@
+import 'core-js'
+import 'regenerator-runtime/runtime'
+
 import Hapi from '@hapi/hapi'
 import inert from '@hapi/inert'
 
 import render from './render'
 
-export const createServer = async () => {
+const createServer = async () => {
   /**
    * BASIC SERVER
    */
@@ -39,3 +42,20 @@ export const createServer = async () => {
 
   return server
 }
+
+async function startServer() {
+  try {
+    const server = await createServer()
+    await server.start()
+    console.log(
+      `Server running at: ${server.info.uri}`,
+      '\n- Open base: http://localhost:8000/base',
+      '\n- Open client: http://localhost:8000/client1',
+    )
+  } catch (err) {
+    console.error(err)
+    process.exit(1)
+  }
+}
+
+startServer()
