+++ packages/example/tsconfig.json
@@ -1,0 +1,23 @@
+/**
+ * https://stackoverflow.com/a/59776114/1238150
+ * https://github.com/dan-kez/lerna-webpack-example/blob/master/tsconfig.settings.json
+ * https://github.com/dan-kez/lerna-webpack-example/blob/master/packages/router/tsconfig.json
+ */
+{
+  /* inspired by https://github.com/marmelab/react-admin/blob/next/packages/ra-data-graphcool/tsconfig.json */
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "declaration": false,
+    "declarationMap": false,
+    "noImplicitAny": false,
+    /* temporary workaround (https://stackoverflow.com/a/52404148/1238150) */
+    "skipLibCheck": true
+  },
+  "include": ["src"],
+  "references": [
+    {
+      "path": "../react-component-catalog"
+    }
+  ]
+}
