+++ packages/react-component-catalog/package.json
@@ -1,0 +1,141 @@
+{
+  "name": "react-component-catalog",
+  "author": "Stefan Natter <https://natterstefan.me>",
+  "license": "Apache-2.0",
+  "version": "2.0.0-beta.0",
+  "description": "Manage a catalog (registry) of react-components easily.",
+  "keywords": [
+    "react-components",
+    "react-registry",
+    "react-catalog",
+    "registry",
+    "catalog",
+    "components",
+    "react"
+  ],
+  "main": "lib/index.js",
+  "module": "esm/index.js",
+  "types": "esm/index.d.ts",
+  "files": [
+    "dist",
+    "es",
+    "esm",
+    "lib"
+  ],
+  "scripts": {
+    "build": " yarn run build-types && yarn run build-cjs && yarn run build-es && yarn run build-esm && yarn run build-umd",
+    "build-types": "yarn run build-cjs-types && yarn run build-es-types && yarn run build-esm-types",
+    "build-cjs": "BABEL_ENV=cjs babel src --ignore **/*.test.tsx,**/*.test.ts --out-dir lib --extensions '.ts,.tsx'",
+    "build-cjs-types": "tsc -p tsconfig.build.json --outDir lib --module commonjs --target es5",
+    "build-es": "BABEL_ENV=es babel src --ignore **/*.test.tsx,**/*.test.ts --out-dir es --extensions '.ts,.tsx'",
+    "build-es-types": "tsc -p tsconfig.build.json --outDir es --module es2015 --target es2015",
+    "build-esm": "BABEL_ENV=esm babel src --ignore **/*.test.tsx,**/*.test.ts --out-dir esm --extensions '.ts,.tsx'",
+    "build-esm-types": "tsc -p tsconfig.build.json --outDir esm --module es2015 --target es5",
+    "build-umd": "BABEL_ENV=umd webpack --mode=production",
+    "coverage": "jest --coverage",
+    "coveralls": "jest --coverage && cat ./coverage/lcov.info | coveralls",
+    "lint": "eslint --cache 'src/**/*.{ts,tsx}' --quiet",
+    "prebuild": "rimraf dist && rimraf es && rimraf esm && rimraf lib",
+    "prepublishOnly": "yarn prerelease",
+    "prerelease": "yarn build && yarn size && yarn test",
+    "release": "HUSKY_SKIP_HOOKS=1 standard-version",
+    "size": "size-limit",
+    "test": "jest --detectOpenHandles && yarn test-types",
+    "test-types": "tsc --noEmit",
+    "watch": "BABEL_ENV=esm babel --watch src --ignore **/@types/*,**/*.test.tsx,**/*.test.ts --out-dir esm --extensions '.ts,.tsx'",
+    "watch-test": "jest --watch"
+  },
+  "dependencies": {
+    "@types/hoist-non-react-statics": "^3.3.1",
+    "hoist-non-react-statics": "^3.3.2"
+  },
+  "peerDependencies": {
+    "react": ">=16.8.3 <17"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged",
+      "prepare-commit-msg": "exec < /dev/tty && git cz --hook"
+    }
+  },
+  "lint-staged": {
+    "*.js": [
+      "yarn run lint",
+      "prettier --write",
+      "git update-index --again",
+      "jest --findRelatedTests"
+    ]
+  },
+  "size-limit": [
+    {
+      "limit": "13 KB",
+      "path": "dist/index.js",
+      "config": "./webpack.config.js",
+      "ignore": [
+        "react",
+        "react-dom"
+      ]
+    },
+    {
+      "limit": "6 KB",
+      "path": "lib/**/*.js"
+    },
+    {
+      "limit": "6 KB",
+      "path": "es/**/*.js"
+    },
+    {
+      "limit": "6 KB",
+      "path": "esm/**/*.js"
+    }
+  ],
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
+  },
+  "standard-version": {
+    "changelogHeader": "# React-Component-Catalog Changelog\n\nAll notable changes to this project will be documented here. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).\n\n",
+    "types": [
+      {
+        "type": "feat",
+        "section": "Features"
+      },
+      {
+        "type": "fix",
+        "section": "Fixes"
+      },
+      {
+        "type": "chore",
+        "hidden": true
+      },
+      {
+        "type": "docs",
+        "hidden": true
+      },
+      {
+        "type": "style",
+        "hidden": true
+      },
+      {
+        "type": "refactor",
+        "hidden": true
+      },
+      {
+        "type": "perf",
+        "hidden": true
+      },
+      {
+        "type": "test",
+        "hidden": true
+      }
+    ]
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/natterstefan/react-component-catalog.git"
+  },
+  "bugs": {
+    "url": "https://github.com/natterstefan/react-component-catalog/issues"
+  }
+}
