+++ packages/react-component-catalog/src/components/__tests__/catalog-component.test.tsx
