+++ packages/react-component-catalog/src/components/with-catalog.tsx
